Introduce compat layer for Window and Rolling
