+++ modin/_compat/pandas_api/abc/window.py
@@ -11,18 +11,20 @@
 # ANY KIND, either express or implied. See the License for the specific language
 # governing permissions and limitations under the License.
 
-from .base import Python36CompatibleBasePandasDataset
-from .dataframe import Python36CompatibleDataFrame
-from .series import Python36CompatibleSeries
-from .groupby import Python36CompatibleDataFrameGroupBy, Python36CompatibleSeriesGroupBy
-from .window import Python36CompatibleWindow, Python36CompatibleRolling
+"""Module for interface for class compatibility layer for window objects."""
 
-__all__ = [
-    "Python36CompatibleBasePandasDataset",
-    "Python36CompatibleDataFrame",
-    "Python36CompatibleSeries",
-    "Python36CompatibleDataFrameGroupBy",
-    "Python36CompatibleSeriesGroupBy",
-    "Python36CompatibleWindow",
-    "Python36CompatibleRolling",
-]
+from modin.logging.class_logger import ClassLogger
+
+
+class BaseCompatibleWindow(ClassLogger):
+    """Interface for class compatibility layer for Window."""
+
+    def _init(self, dataframe, window_args, axis):
+        pass
+
+
+class BaseCompatibleRolling(ClassLogger):
+    """Interface for class compatibility layer for Rolling."""
+
+    def _init(self, dataframe, rolling_args, axis):
+        pass
