+++ modin/_compat/pandas_api/latest/__init__.py
@@ -15,6 +15,7 @@ from .base import LatestCompatibleBasePandasDataset
 from .dataframe import LatestCompatibleDataFrame
 from .series import LatestCompatibleSeries
 from .groupby import LatestCompatibleDataFrameGroupBy, LatestCompatibleSeriesGroupBy
+from .window import LatestCompatibleWindow, LatestCompatibleRolling
 
 __all__ = [
     "LatestCompatibleBasePandasDataset",
@@ -22,4 +23,6 @@ __all__ = [
     "LatestCompatibleSeries",
     "LatestCompatibleDataFrameGroupBy",
     "LatestCompatibleSeriesGroupBy",
+    "LatestCompatibleWindow",
+    "LatestCompatibleRolling",
 ]
