+++ modin/pandas/window.py
@@ -15,38 +15,19 @@
 
 from typing import Optional
 import pandas.core.window.rolling
+from pandas.core.dtypes.common import is_list_like
+
 from modin.utils import _inherit_docstrings
 from modin.logging import ClassLogger
-
-from pandas.core.dtypes.common import is_list_like
+from modin._compat.pandas_api.classes import WindowCompat, RollingCompat
 
 
 @_inherit_docstrings(pandas.core.window.rolling.Window)
-class Window(ClassLogger):
-    def __init__(
-        self,
-        dataframe,
-        window,
-        min_periods=None,
-        center=False,
-        win_type=None,
-        on=None,
-        axis=0,
-        closed=None,
-        method="single",
-    ):
+class Window(WindowCompat):
+    def _init(self, dataframe, window_args, axis):
         self._dataframe = dataframe
         self._query_compiler = dataframe._query_compiler
-        self.window_args = [
-            window,
-            min_periods,
-            center,
-            win_type,
-            on,
-            axis,
-            closed,
-            method,
-        ]
+        self.window_args = window_args
         self.axis = axis
 
     def mean(self, *args, **kwargs):
@@ -82,31 +63,11 @@ class Window(ClassLogger):
     pandas.core.window.rolling.Rolling,
     excluded=[pandas.core.window.rolling.Rolling.__init__],
 )
-class Rolling(ClassLogger):
-    def __init__(
-        self,
-        dataframe,
-        window,
-        min_periods=None,
-        center=False,
-        win_type=None,
-        on=None,
-        axis=0,
-        closed=None,
-        method="single",
-    ):
+class Rolling(RollingCompat):
+    def _init(self, dataframe, rolling_args, axis):
         self._dataframe = dataframe
         self._query_compiler = dataframe._query_compiler
-        self.rolling_args = [
-            window,
-            min_periods,
-            center,
-            win_type,
-            on,
-            axis,
-            closed,
-            method,
-        ]
+        self.rolling_args = rolling_args
         self.axis = axis
 
     def count(self):
